Add files via upload
The purpose of this project was to compare the time difference between uploading large amounts of data to a database via one singular file versus four files, using multi-threading. 
**Important Notes**
The csv_creator.py and csv_splitter.py are two programs I wrote to create and setup arbitrary data to use for my main program, globitech_proj.py. I removed my Postgresql username, password, and database name, so changes to run this program are necessary. The host is the loopback address, 127.0. 0.1, and the port is the database's port.